Rename core.py -> case.py and fix imports
